Registers and memory will be signed after all

+++ src/MainWindow.cpp
@@ -16,11 +16,11 @@ void MainWindow::updateUi() {
     for (int address = 0; address < Processor::MEM_SIZE; ++address) {
         int row = address >> 5;
         int col = address & ((1 << 5) - 1);
-        QString text = QString("%1").arg(pro.memory[address], 2, 16, QLatin1Char('0'));
+        QString text = QString("%1").arg(static_cast<uint8_t>(pro.memory[address]), 2, 16, QLatin1Char('0'));
         memoryTable->setItem(row, col, new QTableWidgetItem(text));
     }
     for (int reg = 0; reg < Processor::REG_COUNT; ++reg) {
-        getReg(reg)->setValue(pro.registers[reg]);
+        getReg(reg)->setValue(static_cast<uint8_t>(pro.registers[reg]));
     }
 }
